Improvement

+++ database/migrations/2025_09_08_065902_create_subjects_tabel.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('code')->unique();
-            $table->decimal('kkm');
+            $table->unsignedTinyInteger('kkm');
             $table->foreignIdFor(GroupSubject::class);
             $table->timestamps();
         });
